security: a split bundle is one archive, and its manifest is a stranger's (audit No. 8)
restore_archive_set had two independent defects, both reachable from a bundle
someone else produced — which is what all three of its callers (verify, handoff,
the desktop bundle viewer) are looking at.

The extraction budget restarted for every part. Each call took a fresh
ExtractLimits::default(), so a 50-part set could expand to 400 GiB with every
individual call comfortably within budget: the zip-bomb ceiling added earlier
simply did not apply to the composite form. ExtractBudget now carries the
remaining allowance across parts, and ExtractLimits is its constructor. The
public signatures are unchanged apart from the extracted count, which becomes
u64 with saturating_add — a member count is bounded by the budget, but the
arithmetic should not be what decides that.

The part names came from ARCHIVE_SET_MANIFEST.json and were joined to the set
directory unchecked. Path::join with an absolute path discards the base
entirely, so an entry naming C:/Users/victim/secret.zip had this function open
that file and unpack it into the destination; "../../x.zip" reached the same
place. Each name now goes through safe_member_target — the check this same file
already applied to the names of members inside an archive. A manifest may also
claim no more than 4096 parts, checked before a single file is opened.

Tests: an absolute part name, a "../" part name, five parts that each fit the
budget and together do not, a manifest claiming too many parts, and the
ordinary set still restoring.

+++ crates/codepack-archive/tests/round_trip.rs
@@ -137,7 +137,7 @@ fn restore_archive_set_round_trip_reproduces_every_file() {
 
     let destination = temp.path().join("restored");
     let extracted = restore_archive_set(&output_dir, &destination).expect("restore");
-    assert_eq!(extracted, result.file_count);
+    assert_eq!(extracted, u64::from(result.file_count));
 
     for relative in [
         "reports/x1.json",
